node driver public API
